AUT-3487: Return appropriate value when mfa method does not exist on a user
This will allow the handler to return the appropriate response (a 404)

+++ shared/src/main/java/uk/gov/di/authentication/shared/services/mfa/DynamoMfaMethodsService.java
@@ -89,6 +89,10 @@ public class DynamoMfaMethodsService implements MfaMethodsService {
             return Either.left(MfaDeleteFailureReason.CANNOT_DELETE_DEFAULT_METHOD);
         }
 
+        if (maybeMethodToDelete.isEmpty()) {
+            return Either.left(MfaDeleteFailureReason.MFA_METHOD_WITH_IDENTIFIER_DOES_NOT_EXIST);
+        }
+
         dynamoService.deleteMfaMethodByIdentifier(email, mfaIdentifier);
         return Either.right(mfaIdentifier);
     }
